Moved time parsing to its own file
